refactor(sc_data): extract the object store the parsed tree mirrors through
ParsedStore holds two things at once: where the sc_data bucket is and how to
move many small objects in and out of a local tree, and on top of that the
mirror semantics the parsed tree needs. Only the second half is about the
parsed tree, and a raw export needs the first half without the mirror.

Moves the settings, the client, the parallel transfer and the listing to
ScData::ObjectStore, which a subclass points at a prefix and a local root. No
behaviour change: the listing now hands back the ETag and the size as a
RemoteObject rather than the ETag alone, so a caller can tell a multipart
upload apart, and ParsedStore reads .etag where it compared the bare value.
