+++ app/lib/sc_data/parsed_store.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "aws-sdk-s3"
-
 module ScData
   # Mirrors one parsed environment tree between `data/sc_data/parsed/<env>` and
   # object storage.
@@ -15,59 +13,17 @@ module ScData
   # carries is removed from the target. `ModelsLoader#update_in_game_flag` reads
   # a model's in-game-ness off file existence, so a leftover file from an
   # earlier build would keep a retired ship flying.
-  class ParsedStore
+  class ParsedStore < ObjectStore
     PREFIX = "parsed"
-
-    # Both trees are ~15k small objects, so the wall clock is round-trip bound
-    # rather than bandwidth bound. Kept modest to stay well inside provider
-    # rate limits on the loaders container.
-    CONCURRENCY = 16
-
-    class NotConfigured < StandardError; end
-
-    class MissingTree < StandardError; end
 
     class VersionMismatch < StandardError; end
 
     attr_reader :environment
 
-    # The parsed tree sits in a different bucket from ActiveStorage's, but on the
-    # same account and endpoint -- and object-storage keys here are
-    # project-scoped, so the pair already deployed for `config/storage.yml`
-    # reaches this bucket too.
-    #
-    # Everything but the bucket therefore falls back to that existing config, so
-    # the only value production carries for this is the one that actually
-    # differs. Each `sc_data_s3` key stays an override, for a separate account
-    # later or for running the CLI outside 1Password.
-    def self.settings
-      creds = Rails.application.credentials
-
-      {
-        endpoint: Rails.app.creds.option(:sc_data_s3, :endpoint) || storage_endpoint(creds),
-        access_key_id: Rails.app.creds.option(:sc_data_s3, :access_key_id) || creds.s3_access_key,
-        secret_access_key: Rails.app.creds.option(:sc_data_s3, :secret_access_key) || creds.s3_secret_key,
-        bucket: Rails.app.creds.option(:sc_data_s3, :bucket)
-      }
-    end
-
-    # `config/storage.yml` stores the scheme apart from the host, so it has to be
-    # rejoined here. Nil rather than a bare "://" when either half is missing, so
-    # `configured?` reports it as absent instead of building a broken client.
-    def self.storage_endpoint(creds)
-      return if creds.s3_protocol.blank? || creds.s3_endpoint.blank?
-
-      "#{creds.s3_protocol}://#{creds.s3_endpoint}"
-    end
-
-    def self.configured?(settings = self.settings)
-      settings.values.all?(&:present?)
-    end
-
     def initialize(environment, client: nil, settings: nil)
+      super(client:, settings:)
+
       @environment = environment.to_s
-      @client = client
-      @provided_settings = settings
     end
 
     # Local -> bucket. Returns the counts of what moved.
@@ -82,7 +38,7 @@ module ScData
 
       remote = remote_objects
 
-      uploaded = each_in_parallel(local.keys.reject { |path| remote[path] == local[path] }) do |path|
+      uploaded = each_in_parallel(local.keys.reject { |path| remote[path]&.etag == local[path] }) do |path|
         File.open(local_path(path), "rb") do |io|
           client.put_object(bucket:, key: key_for(path), body: io)
         end
@@ -106,10 +62,8 @@ module ScData
 
       local = local_files
 
-      downloaded = each_in_parallel(remote.keys.reject { |path| local[path] == remote[path] }) do |path|
-        target = local_path(path)
-        FileUtils.mkdir_p(File.dirname(target))
-        client.get_object(bucket:, key: key_for(path), response_target: target)
+      downloaded = each_in_parallel(remote.keys.reject { |path| local[path] == remote[path].etag }) do |path|
+        download(path)
       end
 
       removed = each_in_parallel(local.keys - remote.keys) do |path|
@@ -152,16 +106,9 @@ module ScData
       "#{PREFIX}/#{environment}"
     end
 
-    private def key_for(path)
-      "#{prefix}/#{path}"
-    end
-
-    private def local_path(path)
-      File.join(local_root, path)
-    end
-
-    # Relative path => MD5 hex, matching the shape of `remote_objects` so the
-    # two can be diffed directly.
+    # Relative path => MD5 hex. Every object in this tree is written
+    # single-part, so a listing's ETag is the plain MD5 of the content and the
+    # two sides can be diffed directly.
     private def local_files
       return {} unless File.directory?(local_root)
 
@@ -172,20 +119,6 @@ module ScData
 
         files[path] = Digest::MD5.file(full).hexdigest
       end
-    end
-
-    # Relative path => ETag. Every object here is written single-part, so the
-    # ETag is the plain MD5 of the content and comparable to a local digest.
-    private def remote_objects
-      objects = {}
-
-      client.list_objects_v2(bucket:, prefix: "#{prefix}/").each do |page|
-        page.contents.each do |object|
-          objects[object.key.delete_prefix("#{prefix}/")] = object.etag.delete('"')
-        end
-      end
-
-      objects
     end
 
     # A directory the mirror emptied is not an object anywhere, so nothing
@@ -201,59 +134,6 @@ module ScData
 
           Dir.rmdir(full) if File.directory?(full) && Dir.empty?(full)
         end
-    end
-
-    private def each_in_parallel(paths)
-      return 0 if paths.empty?
-
-      queue = Queue.new
-      paths.each { |path| queue << path }
-      queue.close # so a drained `pop` returns nil instead of blocking forever
-
-      errors = Queue.new
-
-      Array.new([CONCURRENCY, paths.size].min) do
-        Thread.new do
-          while (path = queue.pop)
-            begin
-              yield path
-            rescue => e
-              errors << "#{path}: #{e.message}"
-            end
-          end
-        end
-      end.each(&:join)
-
-      raise errors.pop unless errors.empty?
-
-      paths.size
-    end
-
-    private def bucket
-      settings.fetch(:bucket)
-    end
-
-    private def settings
-      @settings ||= begin
-        settings = @provided_settings || self.class.settings
-
-        unless self.class.configured?(settings)
-          missing = settings.select { |_, value| value.blank? }.keys
-
-          raise NotConfigured, "missing sc_data_s3 settings: #{missing.join(", ")}"
-        end
-
-        settings
-      end
-    end
-
-    private def client
-      @client ||= Aws::S3::Client.new(
-        endpoint: settings.fetch(:endpoint),
-        access_key_id: settings.fetch(:access_key_id),
-        secret_access_key: settings.fetch(:secret_access_key),
-        region: "unused"
-      )
     end
   end
 end
